[QMS-342] Read GPX files as POIs
Refactor IPoiFile to allow easy adding of new poi files
