+++ src/qmapshack/poi/CPoiItem.cpp
@@ -1,0 +1,25 @@
+/**********************************************************************************************
+    Copyright (C) 2017 Oliver Eichler <oliver.eichler@gmx.de>
+    Copyright (C) 2022 Henri Hornburg <pingurus@t-online.de>
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+**********************************************************************************************/
+
+#include "poi/CPoiItem.h"
+
+uint CPoiItem::getKey() const
+{
+    return qHash(*this, 42);
+}
